requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
playwright>=1.40.0
openai>=1.30.0
reportlab>=4.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.25.0
jinja2>=3.1.0
